requirements.txt
chromadb
python-dotenv
openai
gradio
langchain
langchain-openai